backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.0
python-dotenv==1.0.1
stripe==8.6.0
ccxt==4.2.35
redis==5.0.1
rq==1.15.1
httpx==0.27.0
python-multipart==0.0.9
numpy==1.26.4
cryptography==42.0.5
